Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -17,7 +17,7 @@ int main(){
         break;
         case '/':
         if (b!=0){
-            printf("%d",(float)a/b);
+            printf("%d",a/b);
         }
         else{
             printf("error");
